Write new ChallongeAuthorization class
Rather than having ChallongeAuthorization represent the file itself as well as the web client making requests, I'm thinking of separating the auth file management to its own dedicated class.  This new class extends AbstractMap to make expressing this as a JSON string (for updating the refresh token) really simple.  Also wrote a new PermissionsScopeException to be thrown when an API method is called when the required scope was not given with the grant.