Cleaned up ChallongeAuthorization
Rather than having ChallongeAuthorization extend AbstractSet and deal with having to create or modify Maps for saving a new token or creating the request body to request for a new token, there is now a simple ImmutableMap class that is used to expose both a public requestRequestBody map and house a private fileSaveBody.  Since they're both ImmutableMaps constructed with references to the same Entry objects, calling refresh_token.setValue() reflects in both!