(probably) final attempt at TournamentOptions
TournamentOptions is a base class of which DoubleEliminationOptions etc. extend.  All of the subclasses pretty much just call the super constructor while maintaining a unique type for TournamentType enums to specify as a requirement.  Thus, depending on which TournamentType you want to set a ChallongeTournament to, a specific TournamentOptions will be required.