Change ChallongeTournament update method
Rather than accepting each value to be updated as a parameter (which is cumbersome, considering you need to send the entire tournament back in the body), there is now a toAttributes method which packs the tournament back into an attributes JSONObject which can be sent in the update method.  Downside to this is that an error from the API will result in a desync and possible broken tournament object, though this has not been observed.