Single Elimination support, fixes
SingleEliminationOptions is a new class which exists solely to differentiate tournament types.  DoubleEliminationOptions now properly parses grandFinalsModifier which can sometimes be null, which defaults to GrandFinalsModifier.NONE.    TournamentOptions.getOptions is no longer public since it should never be accessed outside of the package to begin with.  If a call to ChallongeTournament.update() fails for any reason, it will refresh its fields from the API (hoping this doesn't cause any issues down the line as parity between local and online model should be maintained as much as possible, and all fields are immutable).