Fix for GrandFinalsModifier.NONE
Even though API documentation says an empty string can be used to represent the default none grand_finals_modifier, the API returns an error.  The default is actually represented as null when returned from the API but null can't be passed through the update endpoint either.  The only way to represent no grand_finals_modifier is by not including it as an entry.  Thus, if GrandFinalsModifier.NONE is passed into DoubleEliminationOptions constructor, it get passed as null into the ImmutableMap which then filters it out (which should be fine, since ImmutableMap's functionality in this package depends on it always having Entry objects ONLY)